[feature] cache 없이 키워드&배달위치 기준 검색 기능

+++ domain/exposure-common-domain/src/main/java/woowa/team4/bff/domain/RestaurantSummary.java
@@ -14,6 +14,6 @@ public class RestaurantSummary {
     private UUID restaurantUuid;
     private String restaurantName;
     private String restaurantThumbnailUrl;
-    private Integer reviewCount;
+    private Long reviewCount;
     private Double rating;
 }
